requirements.txt
# install of metalad itself
.
nose
coverage
sphinx>=1.7.8
sphinx-rtd-theme
python-dateutil
pyyaml
